requirements.txt
joblib==1.4.2
lightgbm==4.6.0
numpy==2.2.5
opencv_contrib_python==4.11.0.86
opencv_python==4.11.0.86
pandas==2.2.3
Pillow==11.2.1
scikit_learn==1.6.1
skimage==0.0
torch==2.2.1+cu121
torchvision==0.17.1+cu121
tqdm==4.67.1
